+++ cedar-policy-symcc/src/symcc/solver_pool.rs
@@ -266,6 +266,15 @@ impl Solver for PooledSolver {
         }
     }
 
+    async fn enable_models(&mut self) -> Result<(), SolverError> {
+        // The `PooledSolver` implementation doesn't need to do anything other
+        // than set the appropriate SMTLib option.
+        self.smtlib_input()
+            .set_option("produce-models", "true")
+            .await
+            .map_err(Into::into)
+    }
+
     async fn check_sat(&mut self) -> Result<Decision, SolverError> {
         let solver = self
             .solver