+++ cedar-policy-symcc/CHANGELOG.md
@@ -16,6 +16,9 @@ Cedar Language Version: TBD
 
 - Refactored `Solver` trait to include `check_sat_with_model()` instead of `get_model()`,
 to make the API easier to use with fewer implicit ordering constraints (#2190)
+- Added `enable_models()` as a required method for the `Solver` trait, to allow
+some implementors of `Solver` to perform other initialization/configuration as
+required when models are being enabled (#2192)
 
 ### Fixed